abstract out createObservable
